Supports tarballs and better downtime deploys
*.tar & *.tar.gz archives are now supported

Changes the deployment strategy. Now the downloader will pull down the
archive and extract it into a directory that is the basename of the
archive. Afterwards a new symlink between that new directory and
`current` is established. The app server is reading from `current` and
will bounce when it changes.

This allows for shorter downtime as previously the downloader would tear
down the app servr, remove the old code directory and replace it with
the new code. Then the app server was brought back up. Depending upon
the app size this could represent a few seconds worth of downtime.

In addition, in the event of wanting to fallback to the non-fastboot app
when the app server is not running, by deleting the old code made this
impossible. With this new strategy while the app server is offline you
can have Nginx fallback to the `index.html` in the code directory. Or in
the event of an app server crash during the download event the
non-fastboot version is still available for use.